--- src/Components/Home.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Home = () => {
-  return (
-    <div>Home Estas dentro</div>
-  )
-}
-
-export default Home
