+++ src/Components/Home/Home.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Home = () => {
+  return (
+    <div>Home Estas dentro</div>
+  )
+}
+
+export default Home
