+++ src/Components/Register/register.css
@@ -1,0 +1,3 @@
+.but {
+    font-size: 1.2rem;
+  }
