--- src/register.css
@@ -1,3 +1,0 @@
-.but {
-    font-size: 1.2rem;
-  }
